Implement constructor and print()
Canvas says that we have a file named contactlist.cpp, but it also
alludes to us implementing a phonebook class. All of the methods in
contact.hpp are defined and empty (meaning that they end in {} and not
;), so it is impossible to override them and make them actually do
something. What I've done is put a class called PhoneBook in
contactList.cpp that extends the List class in
contact.hpp. Unfortunately, head is private in the List class, so
PhoneBook needs it's own head node. This means we're literally redoing
everything that's in the List class. This isn't very elegant, but I
don't see any other way to follow all of the directions. If you have
any ideas on how to improve this, they are welcome.

Also, @Maren, I think we have to put both of our work in this
file. I've already added you as an author, and you're welcome to add
your methods to the class definition.