fix compile errors in insert and remove
Fix typo in insert declaration

insert is declared within the class definition, so we need to replace
PhoneBook::insert with insert

Fix declaration of local variables

It looks like the end of a line accidently got deleted. Now
nodeToInsert is correctly declared and initialized to contain the
contact passed to insert

Add a name to the Contact passed to insert

insert was declared as insert(const Contact), but now it's
insert(const Contact contact)

Replace all calls to Node.lastName with node->data.lastName

Node doesn't contain any data about the contacts name. In order to
acces it you have to first refer to the Node's memember variable named
data with Node->data.

Fix error with insert determining whether currentNode is null

insert was checking if currentNode->data.lastName was null instead of
currentNode

Fix declaration of remove

The single parameter for remove didn't have a type. It was just
declared as value. I fixed it to be a const Contact, and I renamed
value to contact to match the parameter in insert. It is now
remove(const Contact contact)